Move ArticleXML methods not requiring self-use over to a build module.
